# Remove interim slice in `IsDisjoint`